Update README.md
Original text:
"This Wacom Screen Rotation Script will be supporting for more Tablet Computers automatically. The goal is to integral with caja-actions or other ease-to-use way for users to rotation the Linux Screen if the accelerometer is not detected by Linux."

Revised to:
"This script will support more Tablet Computers automatically. The goal is to integrate the Caja-actions or other ways for users to rotate the Linux Screen easily if the accelerometer is not detected by Linux."